requirements.txt
Django~=5.1
pillow~=10.4.0
python-dotenv==1.0.1